tests/qr-code-testing/requirements.txt
qrcode>=7.4.2
Pillow>=10.0.0
requests>=2.31.0
pytest>=7.4.0
playwright>=1.40.0
